fixup: consolidate types, constants
