--- libs/hooks/open-telemetry/src/lib/constants.ts
@@ -1,5 +1,0 @@
-export const FEATURE_FLAG = 'feature_flag';
-export const ACTIVE_COUNT_NAME = `${FEATURE_FLAG}.evaluation_active_count`;
-export const REQUESTS_TOTAL_NAME = `${FEATURE_FLAG}.evaluation_requests_total`;
-export const SUCCESS_TOTAL_NAME = `${FEATURE_FLAG}.evaluation_success_total`;
-export const ERROR_TOTAL_NAME = `${FEATURE_FLAG}.evaluation_error_total`;
